+++ example/comparators_example.dart
@@ -1,5 +1,5 @@
+import 'package:collection/collection.dart' show ComparatorExtension;
 import 'package:comparators/comparators.dart';
-import 'package:comparators/extensions.dart';
 
 final List<User> users = [
   const User(
@@ -41,41 +41,47 @@ final List<User> users = [
 ];
 
 void main() {
-  _printUsers();
+  var usersCopy = [...users];
+  void reset() => usersCopy = [...users];
+  void printUsers({String trailing = '\n'}) =>
+      print('${usersCopy.join('\n')}$trailing');
+
+  printUsers();
 
   // this will sort the list by the username field of the User object
-  users.sort(compare((u) => u.username));
+  usersCopy.sort(compare((u) => u.username));
 
-  _printUsers();
-
-  // this will sort the users by their username
-  // before comparing the usernames will be transformed with the provided transform
-  // in this case, it will lowercase the names to do a case insensitive comparison
-  users.sort(
-    compareTransformed<User, String>(
-        (u) => u.username, (name) => name.toLowerCase()),
-  );
-
-  _printUsers();
+  printUsers();
+  reset();
 
   // this will sort the users by their activity first, then by their email,
   // and then by their username
-  users.sort(
+  // * `inverse` is from the `package:collection`
+  usersCopy.sort(compareSequentially([
     // the users which active is set to true will come first in the list
-    compareBool<User>((u) => u.isActive).reversed.then(
-          // then users will be sorted by their email field
-          compare<User>((u) => u.email).then(
-            // and then by their username
-            compare<User>((u) => u.username),
+    // note that using `inverse` requires explicit type in the comparator
+    compareBool((User u) => u.isActive).inverse,
+    // then users will be sorted by their email field
+    compare((u) => u.email),
+    // and then by their username
+    compare((u) => u.username),
+  ]));
+
+  printUsers();
+  reset();
+
+  // alternatively, it is possible to achieve the same result using `then` and `inverse` extension methods from the
+  // `package:collection`, but it is necessary to write types explicitly in most cases
+  usersCopy.sort(
+    compareBool((User u) => u.isActive).inverse.then(
+          compare((User u) => u.email).then(
+            compare((User u) => u.username),
           ),
         ),
   );
 
-  _printUsers(trailing: '');
+  printUsers(trailing: '');
 }
-
-void _printUsers({String trailing = '\n'}) =>
-    print('${users.join('\n')}$trailing');
 
 /// A class representing a user with an id, username, email and activity status.
 class User {