+++ lib/src/extensions/comparator_chaining.dart
@@ -1,4 +1,4 @@
-/// A [Comparator] extension adding [then] method to allow chaining the
+/// A [Comparator] extension adding `then` method to allow chaining the
 /// comparators together.
 extension ComparatorChaining<T> on Comparator<T> {
   /// Returns a new [Comparator] combining this one and the given
@@ -6,6 +6,8 @@ extension ComparatorChaining<T> on Comparator<T> {
   ///
   /// Firstly, this comparator will be used, and if the result is 0, the
   /// [nextComparator] will be used.
+  @Deprecated('It duplicates the functionality of \'then\' '
+      'from the \'package:collection\' and will be removed in v2.0.0')
   Comparator<T> then(Comparator<T> nextComparator) => (a, b) {
         final firstComparison = call(a, b);