+++ lib/src/field_comparators.dart
@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart' show ComparatorExtension;
 import 'package:comparators/src/util.dart';
 
 import 'typedefs.dart';
@@ -23,7 +24,14 @@ import 'typedefs.dart';
 /// );
 /// // in this package, there is `compareBool` function which does the same same as in this example
 /// ```
+@Deprecated('It will be made private starting from v2.0.0')
 Comparator<T> compareTransformed<T, R>(
+  FieldExtractor<T, R> fieldExtractor,
+  ComparableTransformer<R, Comparable> comparableTransformer,
+) =>
+    _compareTransformed(fieldExtractor, comparableTransformer);
+
+Comparator<T> _compareTransformed<T, R>(
   FieldExtractor<T, R> fieldExtractor,
   ComparableTransformer<R, Comparable> comparableTransformer,
 ) {
@@ -37,9 +45,6 @@ Comparator<T> compareTransformed<T, R>(
 
 /// Returns a [Comparator] of type [T] comparing by the [Comparable] field of
 /// type [R] extracted with the [fieldExtractor].
-///
-/// It is a shorthand for the [compareTransformed] comparator function for the
-/// fields that are [Comparable] by themselves.
 ///
 /// Example:
 /// ```dart
@@ -58,7 +63,8 @@ Comparator<T> compareTransformed<T, R>(
 /// );
 /// ```
 Comparator<T> compare<T>(FieldExtractor<T, Comparable> fieldExtractor) {
-  return compareTransformed<T, Comparable>(fieldExtractor, identityTransformer);
+  return _compareTransformed<T, Comparable>(
+      fieldExtractor, identityTransformer);
 }
 
 /// Returns a comparator for a boolean field extracted with the given
@@ -67,5 +73,50 @@ Comparator<T> compare<T>(FieldExtractor<T, Comparable> fieldExtractor) {
 /// Internally it will use the integer comparison and the following
 /// transformation: `true => 1, false => 0`.
 Comparator<T> compareBool<T>(FieldExtractor<T, bool> fieldExtractor) {
-  return compareTransformed(fieldExtractor, boolTransformer);
+  return _compareTransformed(fieldExtractor, boolTransformer);
+}
+
+/// Returns a comparator that will compare the values of [T] using the [comparators] in their order in the iterable.
+/// Next comparators are used as tie breakers for the previous ones.
+///
+/// This approach allows Dart to infer the comparator types from the context and does not require providing generics
+/// explicitly for most cases.
+///
+/// Example:
+/// ```dart
+/// // chaining using `then` from the `package:collection`
+/// users.sort(
+///   compare<User>((user) => user.name).then(
+///     compare<User>((user) => user.surname).then(
+///       compare<User>((user) => user.country),
+///     ),
+///   ),
+/// );
+///
+/// // with `compareSequentially`
+/// users.sort(compareSequentially([
+///   compare((user) => user.name),
+///   compare((user) => user.surname),
+///   compare((user) => user.country),
+/// ]));
+///
+/// // using `inverse` from the `package:collection`
+/// users.sort(compareSequentially([
+///   // ...
+///   compare((User user) => user.surname).inverse,
+///   // ...
+/// ]));
+/// ```
+Comparator<T> compareSequentially<T>(Iterable<Comparator<T>> comparators) {
+  if (comparators.isEmpty) {
+    throw StateError('An empty iterable of Comparators cannot be combined');
+  }
+
+  var combined = comparators.first;
+
+  for (final comparator in comparators.skip(1)) {
+    combined = combined.then(comparator);
+  }
+
+  return combined;
 }