+++ test/extensions_test.dart
@@ -11,6 +11,7 @@ import 'utils.dart';
 
 void main() {
   final rand = Random();
+  // ignore: deprecated_member_use_from_same_package
   final comparator = compare<NotComparable>((nc) => nc.intValue).then(
     compareBool<NotComparable>((nc) => nc.boolValue),
   );
@@ -22,13 +23,6 @@ void main() {
         final list = rList(rand);
 
         final listMatcher = list.toList();
-        int matcherComparator(NotComparable a, NotComparable b) {
-          if (a.intValue != b.intValue) return a.intValue.compareTo(b.intValue);
-
-          final aB = a.boolValue ? 1 : 0, bB = b.boolValue ? 1 : 0;
-
-          return aB.compareTo(bB);
-        }
 
         list.sort(comparator);
         listMatcher.sort(matcherComparator);
@@ -43,6 +37,7 @@ void main() {
         final list = rList(rand);
         var listMatcher = list.toList();
 
+        // ignore: deprecated_member_use_from_same_package
         list.sort(comparator.reversed);
         listMatcher = (listMatcher..sort(comparator)).reversed.toList();